r275/sofa-dev : Fully functional scenegraph-based version! Added dynamic scenegraph view in QT GUI. Removed obsolete classes. Converted all demos. Simplified example1 (removed procedural code). Temporarily disabled inertial forces related to moving coordinate systems. One remaining bug with Static Solver... Warning: not tested under Linux yet.
 //partial sync (130/234 changes)//

+++ src/Sofa/Components/Gravity.h
@@ -17,7 +17,7 @@ class GNode;
 
 class Gravity : public Abstract::ContextObject
 {
-    typedef Common::Vec3f Vec3;
+    typedef Common::Vec3d Vec3;
 public:
     Gravity();
     const Vec3&  getGravity() const;
